Refactor sql wire impl for tasks
This cleans things up to add namespace query support.
